# Java/Lab6,7/InternetProvider/output/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
...
paths:
  /service/update:
    put:
      tags:
      - Service controller
      summary: Update service
      description: Changing data of service using service id and new params
      operationId: updateService
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      - name: type
        in: query
        required: true
        schema:
          type: string
      - name: price
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Service updated
          content:
            '*/*':
              schema:
                type: string
  /payment/update:
    put:
      tags:
      - Payment controller
      summary: Update payment
      description: "Changing installation payment using installation id, payment id\
        \ and new params"
      operationId: updateInstallationPayment
      parameters:
      - name: paymentId
        in: query
        required: true
        schema:
          type: string
      - name: date
        in: query
        required: true
        schema:
          type: string
      - name: price
        in: query
        required: true
        schema:
          type: string
      - name: installationId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment updated
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /service/add:
    post:
      tags:
      - Service controller
      summary: Add service
      description: Adding service using type and price
      operationId: addService
      parameters:
      - name: type
        in: query
        required: true
        schema:
          type: string
      - name: price
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Service added
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /payment/add:
    post:
      tags:
      - Payment controller
      summary: Add payment
      description: Adding payment to specified installation using installation id
        and payment params
      operationId: addInstallationPayment
      parameters:
      - name: date
        in: query
        required: true
        schema:
          type: string
      - name: price
        in: query
        required: true
        schema:
          type: string
      - name: installationId
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Payment added
          content:
            '*/*':
              schema:
                type: string
...
  /service/get/{id}:
    get:
      tags:
      - Service controller
      summary: Get service by id
      description: Get service using id
      operationId: getClientData
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Service fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /payment/get/{installation-id}:
    get:
      tags:
      - Payment controller
      summary: Get installation payments
      description: Fetching all installation payments using installation id
      operationId: getInstallationPayments
      parameters:
      - name: installation-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Installation payments fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /client/get/{id}:
    get:
      tags:
      - Client controller
      summary: Get client using id
      description: Fetching data of specified client using id
      operationId: getClientData_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Client fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /service/delete:
    delete:
      tags:
      - Service controller
      summary: Delete service
      description: Deleting service using id
      operationId: deleteService
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Service deleted
          content:
            '*/*':
              schema:
                type: string
  /payment/delete:
    delete:
      tags:
      - Payment controller
      summary: Delete payment
      description: Deleting payment using id
      operationId: deleteInstallationPayment
      parameters:
      - name: paymentId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment deleted
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...